extract test helpers to dedicated file
